+++ src/dialog/dialog.component.ts
@@ -152,25 +152,25 @@ export class IgxDialogComponent implements IToggleView, OnInit, OnDestroy {
         return this._titleId;
     }
 
-    get titleTemplate(): any {
-        if (this.titleTemplateDirective) {
-            return this.titleTemplateDirective.template;
-        }
-        return null;
-    }
+    // get titleTemplate(): any {
+    //     if (this.titleTemplateDirective) {
+    //         return this.titleTemplateDirective.template;
+    //     }
+    //     return null;
+    // }
 
-    get buttonsTemplate(): any {
-        if (this.buttonsTemplateDirective) {
-            return this.buttonsTemplateDirective.template;
-        }
-        return null;
-    }
+    // get buttonsTemplate(): any {
+    //     if (this.buttonsTemplateDirective) {
+    //         return this.buttonsTemplateDirective.template;
+    //     }
+    //     return null;
+    // }
 
-    @ContentChild(forwardRef(() => IgxDialogTitleDirective), { read: IgxDialogTitleDirective })
-    private titleTemplateDirective: IgxDialogTitleDirective;
+    // @ContentChild(forwardRef(() => IgxDialogTitleDirective), { read: IgxDialogTitleDirective })
+    // private titleTemplateDirective: IgxDialogTitleDirective;
 
-    @ContentChild(forwardRef(() => IgxDialogActionsDirective), { read: IgxDialogActionsDirective })
-    private buttonsTemplateDirective: IgxDialogActionsDirective;
+    // @ContentChild(forwardRef(() => IgxDialogActionsDirective), { read: IgxDialogActionsDirective })
+    // private buttonsTemplateDirective: IgxDialogActionsDirective;
 
     constructor(
         private elementRef: ElementRef,